chore: Move multipart upload functions into a separate service
