+++ backend/src/core/services/upload.service.ts
@@ -65,103 +65,6 @@ const extractParamsFromJwt = (
 }
 
 /**
- * Get a presigned url to upload a part for multipart upload.
- */
-const getPresignedPartUrl = async ({
-  s3Key,
-  uploadId,
-  partNumber,
-}: {
-  s3Key: string
-  uploadId: string
-  partNumber: number
-}): Promise<string> => {
-  const params = {
-    Bucket: FILE_STORAGE_BUCKET_NAME,
-    Key: s3Key,
-    PartNumber: partNumber,
-    UploadId: uploadId,
-  }
-  return await s3.getSignedUrlPromise('uploadPart', params)
-}
-
-/**
- * Create a multipart upload on s3 and return the upload id and s3 key for it.
- */
-const startMultipartUpload = async (
-  mimeType: string,
-  partCount: number
-): Promise<{ transactionId: string; presignedUrls: string[] }> => {
-  const s3Key = uuid()
-
-  const params = {
-    Bucket: FILE_STORAGE_BUCKET_NAME,
-    Key: s3Key,
-    ContentType: mimeType,
-  }
-
-  const { UploadId } = await s3.createMultipartUpload(params).promise()
-
-  if (!UploadId) throw new Error('no upload id')
-
-  const transactionId = jwtUtils.sign({
-    uploadId: UploadId,
-    s3Key,
-  })
-
-  const presignedUrlPromises = []
-  for (let partNumber = 1; partNumber <= partCount; partNumber++) {
-    presignedUrlPromises.push(
-      getPresignedPartUrl({ s3Key, uploadId: UploadId, partNumber })
-    )
-  }
-  const presignedUrls = await Promise.all(presignedUrlPromises)
-
-  return {
-    transactionId,
-    presignedUrls,
-  }
-}
-
-/**
- * Complete the multipart upload.
- * Returns the s3Key for the upload.
- */
-const completeMultipartUpload = async ({
-  transactionId,
-  partCount,
-  etags,
-}: {
-  transactionId: string
-  partCount: number
-  etags: Array<string>
-}): Promise<string> => {
-  const parts = []
-  for (let i = 0; i < partCount; i++) {
-    parts.push({
-      ETag: etags[i],
-      PartNumber: i + 1,
-    })
-  }
-  const { s3Key, uploadId } = extractParamsFromJwt(transactionId) as {
-    s3Key: string
-    uploadId: string
-  }
-
-  const params = {
-    Bucket: FILE_STORAGE_BUCKET_NAME,
-    Key: s3Key,
-    MultipartUpload: {
-      Parts: parts,
-    },
-    UploadId: uploadId,
-  }
-
-  await s3.completeMultipartUpload(params).promise()
-  return s3Key
-}
-
-/**
  * On file upload complete, save the transaction id and file name against the campaign so that we can download the file from s3 later
  */
 const replaceCampaignS3Metadata = (
@@ -333,8 +236,6 @@ export const UploadService = {
   /*** S3 API Calls ****/
   getUploadParameters,
   extractParamsFromJwt,
-  startMultipartUpload,
-  completeMultipartUpload,
   /**** Handle S3Key in DB *****/
   replaceCampaignS3Metadata,
   storeS3TempFilename,
